ceph_test_rados_api_list: fix warning
In file included from ./test/librados/TestCase.h:9:0,
                 from test/librados/list.cc:7:
../src/gmock/gtest/include/gtest/gtest.h: In instantiation of ‘testing::Asserti$
../src/gmock/gtest/include/gtest/gtest.h:1484:23:   required from ‘static testi$
test/librados/list.cc:944:5:   required from here
../src/gmock/gtest/include/gtest/gtest.h:1448:16: warning: comparison between s$
   if (expected == actual) {
                ^

+++ src/test/librados/list.cc
@@ -941,7 +941,7 @@ TEST_F(LibRadosListPP, EnumerateObjectsFilterPP) {
     std::vector<ObjectItem> result;
     int r = ioctx.object_list(c, end, 12, filter_bl, &result, &c);
     ASSERT_GE(r, 0);
-    ASSERT_EQ(r, result.size());
+    ASSERT_EQ(r, (int)result.size());
     for (int i = 0; i < r; ++i) {
       auto oid = result[i].oid;
       // We should only see the object that matches the filter
